fix analysis and transformations in presence of invalidate
most of our analysis and transformation where broken if
invalidate occurs in the instruction stream. The reason is
that invalidate has two successors -- one of those lives
in a different context.

we could fix this by enriching all our analysis with
different out_sets for the fall through case as for
the invalidate case. But this would completely defeat the
purpose of having invalidate in the first place: namely
to be able to ignore the deopt case!

the logical conclusion is that we should not analyze
nor optimize the deopt continuations of the invalidate!

after experimenting with several solutions I decided to
introduce a new end_opt instruction. every transformation
has to respect the following invariant: code after the
end_opt instruction is not to be changed.

the advantage is that there is a source-level equivalent
of the split between actual function code we want to
optimize and continuations which should stay unchanged.

this allows us to write down already optimized programs
and send them again through our full optimization pipeline.