fix codemotion
for codemotion to be correct we need to rename the moved definitions!
otherwise it is possible to capture and override unrelated use-def
chains with the same variable name.

BUT: for renaming to be limited to the active program (and not the
deopt landing pads) we need to be able to preserve the names in the
invalidate case.

Therefore invalidate was extended to support renaming:

invalidate exp branch [out_names, ... = in_names, ...]

where in_names are the captured variables and out_names are their
new names in the invalidate branch