fix hoist_assign
There where two things broken with the pass:

1. It might hoist assignments to a place where the rhs
   expression is invalid
2. It might destroy the scope

The first problem occurs for example in:

     var a = true
     branch (a==1) isint cont
    isint:
     a <- a+2
    cont:

If we hoist the assignment to line 2 the addition throws
a type error. The problem is solved by only hoisting assignments
with simple expressions rhs. A better approach would be nice...

The second problem arises for example in:

     branch ... a b
    a:
     var a = 1
     ...
     a <- 2
     ...
     drop a
   goto b

If we create a new variable a_1 inside the branch we need to
drop it before merging the scope with b. This problem is solved
by running a generic fix_scope pass which makes all missing drops
explicit.