fix hoist drop
Hoist drop was broken in the case from examples/scope_err.sou

It was assumed that a split edge is only necessary if the first label
to encounter sees two branches and otherwise a pull is ok.

But e.g. in the case:

        a    b

       /  \  /

      c    d

we would happily assume at `d` that the graph is in split-edge form
(because only one branch is a successor). Then we would pull on `a`
and if that succeeds the push to `b` gets lost.

The workaround is to more strongly normalize the graph, ie. allow
mergeing control flow only through gotos. This fixes the missing push
problem. Secondly split edge is performed ahead of time to avoid having
to do non-local checks for all successors of predecessors of a label.