fix constantfolding type confusion
constant fold would happily convert `a[b]` into `nil[b]`. The main
reason is that our operations where not "typed" and therefore we
had no way of specifying which position should be a variable or an
arbitrary expression.

This commit changes the structure of expressions to syntactically
enforce the correct argument types.